refactor(auto): extract Capstone detail inspection into a pure tested seam
The whole-binary `--auto` candidate finder must ask three
soundness-critical questions of every decoded instruction — is it a
call, does it touch RIP-relative memory, and which direct branch
targets does it name — but Capstone answers them only through a
borrowed `InsnDetail`. Those predicates and the reducer that turns one
detail into owned `CapstoneInstructionFacts` sat inline in `main.rs`,
interleaved with ELF I/O and the run-splitting state machine, so a
group-filter or operand-typing rule could only be reached by building a
whole ELF fixture in memory.

Lift them into `src/capstone_detail.rs` as a pure seam: the module owns
`inspect_capstone_instruction_detail` plus the three
`capstone_detail_*` predicates, and `main.rs` keeps only the ELF
plumbing that calls them. This is the Capstone-facing counterpart to the
pure run-splitting algorithm already in `candidate_windows.rs`, and it
drops ~295 lines from `main.rs`.

The four predicate tests move with the code and gain two new direct
unit tests for `inspect_capstone_instruction_detail` (call and
RIP-relative reductions), which previously had only integration
coverage. Behavior is unchanged; the two integration tests that
embedded a redundant predicate assertion now rely on the module's unit
tests instead.
